+++ src/gui/widgets/dialogs/glaxnimate_window.cpp
@@ -386,7 +386,9 @@ void GlaxnimateWindow::dropEvent(QDropEvent* event)
 {
     auto str = d->drop_event_data(event);
     if ( !str.isEmpty() )
-        document_open(str);
+    {
+        d->drop_file(str);
+    }
 }
 
 void GlaxnimateWindow::switch_tool(tools::Tool* tool)